Allocate all inter-partition temporaries to a single Memory object.
Previous behavior was to allocate each inter-partition temporary
to its own Memory object.  The Android system limits the number
of shared memory objects, which is what our Memory objects represent;
so as long as all inter-partition temporaries have the same lifetime
(which they do in our present implementation) it is better to put
them all in the same Memory object.

Bug: 63905942
Test: mma (userdebug)
      ml/nn/runtime/tests (userdebug)
      (with debug.nn.partition.test 1 and 2;
       no new failures)

Change-Id: Ia0ccf4e47aa93d40db0ff5f496662523622caf08