Add CPU fallback and remove non-plan driver path.
When we do ExecutionPlan-controlled execution, and
there is a failure at some stage of the execution,
fall back to cpu execution for that one stage (if
possible) or for the entire model (otherwise).

When we don't do ExecutionPlan-controlled execution,
we no longer attempt to find a suitable driver for
full-model execution -- instead, we execute the full
model on the cpu.  (Rationale: This avoids having
to implement cpu fallback outside ExecutionPlan-
controlled execution.)

Bug: 63905942
Test: mma (userdebug)
      ml/nn/runtime/tests (userdebug)
      (with debug.nn.partition.test 1,
       no new failures; hand-insert
       errors in the runtime in a few
       places to exercise some of the
       fallback paths, and verified that
       behavior is as expected)

Change-Id: I9ec8fba632f94e04840a786575d338d06fd2dcd3