Generate MobilNet quantized in the test generation script
Generate MobileNet quantized and prevent the test from being
overwritten. It was with a different suffix from the one recognized by
the test geneartor script, and thus effectively it was ignored.

This CL updates the suffix of the test spec for MobileNet. Also, since
the example file is generated elsewhere, we want the generator script
not to overwrite it.

Also, skip MobileNet quantized entirely in VTS as it is known to fail
at this moment.

Test: NeuralNetworkTest on Angler
Test: VtsHalNeuralnetworksV1_0TargetTest on CPU sample driver
Bug: 67394906
Change-Id: Ia944aeaa74e9318014f34cfc2fea6dcbf3f0f1ba