Upgrade 1.0 sample driver to 1.1
This CL updates the 1.0 sample drivers to 1.1. Because V1_1::IDevice
inherits from V1_0::IDevice, the sample driver can still be used as a
1.0 driver. When the NN runtime holds a 1.1 driver, it will always use
its updated *_1_1 function calls; to verify the NN runtime is still
compliant with 1.0 drivers, the VTS framework generates a binary called
'android.hardware.neuralnetworks@1.0-adapter' to fake a 1.0 driver using
a 1.1+ driver.

This change additionally updates RandomPartitioningTest to use the 1.1
version of SampleDriver, and removes driver/Android.bp since it is no
longer needed in the build system.

Bug: 63911257
Test: mm
Test: cts and vts test binaries
Change-Id: I1f98d1329571a4571c1b8d5f532ca04dad5a59ca