Simplify lifetime management of ModelBuilder
ExecutionStep fully owns the ModelBuilder it contains, and only provides
references to the subModel. This change simplifies the lifetime
management to reflect the ownership, removing the unnecessary strong
pointer behavior.

Additionally, this CL removes the unnecessary virtual destructor in the
ModelBuilder class, which was likely an artifact of a previous
implementation.

Test: mm
Test: NeuralNetworksTest
Test: VtsHalNeuralnetworksV1_0TargetTest
Change-Id: I62571d9a61fd8bbc193586756267ad4ec8888e27