Prevent excessive logging of test failures
In the generated tests, whenever the produced output data does not
match the expected output data, each mismatched element is logged.
For tensors with thousands of elements, this can spam logcat and
take several minutes just to finish printing the test failure.

This CL caps the total amount of elementwise error log messages to a
preset number so that at most N error messages will be displayed per
test. It will additionally log how many values were mismatched in a
given test as a summary.

Bug: 73253476
Test: mma
Test: NeuralNetworksTest
Change-Id: I13be298537318bda7beccefc6079ed606df17987