Only export public APIs from libneuralnetworks.so.
As a consequence, we can no longer have NeuralNetworksTest equivalent to
NeuralNetworksTest_static, because some of NeuralNetworksTest* relies on
non-public symbols from the runtime.  Therefore, replace
NeuralNetworksTest with NeuralNetworksTest_shared_partial, which links
against libneuralnetworks.so but contains only the subset of tests that
do not rely on non-public symbols from the runtime.

Bug: 73169876

Test: NeuralNetworksTest_static
Test: NeuralNetworksTest_shared_partial
Test: CtsNNAPITestCases
Change-Id: I0a57315944da338c37d91e5ea1ca2d5f0c19a212