Map execution failure error status from driver through NN API to client.
Previously, we only ensured that we get ANEURALNETWORKS_NO_ERROR iff
ErrorStatus::NONE -- that is, we knew whether or not we have an error,
but not the kind of error we have.

With this change, we try to do a sensible mapping from driver error
kind to API error kind.

Test: NeuralNetworksTest_static

Bug: 77639748

Change-Id: I2d186c11c3fe01150cd0f66d40882590c2c05e9b