Fix slicing.sh for new directory structure
Fix path handling in slicing.sh and helpers to work with versioned test
directories.

Added "set -eu" and a number of other checks to scripts to make them
easier to debug in the future.

Renegerated CTS tests with new scripts. The changes were not due to
script changes in this CL, but due to not always running the scripts
when changing models.

TODO: set up a continuous build for slicing.

Bug: 73560182
Test: ./slicing.sh 2 V1_0/mobilenet_quantized.mod.py
Test: ./runtime/test/specs/generate_test.sh and NeuralNetworksTest_static
Change-Id: I9cc00df1c10fc369410388d63c34689647c5c56b