Sync updated NNAPI doc to HAL -- script
Additionally reformatted text to fit within 80 character line limit for
HIDL HALs.

Additionally removed "of each addition" from "Specifies the activation
to invoke on the result of each addition" because it was was included
regardless of the operation. As one example, it was included in CONV_2D,
which (although it includes a bias add) was confusing and unclear.

Bug: 72650109
Test: None, only documentation change
Change-Id: I779654aa30e1dec1f376ca173257e839fc68dc64