Document, support, and test unspecified dimension.
A tensor operand type with some number of unspecified dimensions is
represented by setting each unspecified dimension to 0. A tensor operand
type of unspecified rank is represented by setting dimensionCount to 0
and dimensions to NULL.

Tensor with unspecified dimensions can not have value and can not be as
model input at execution time, but is allowed for output. Enable checks
and tests for it.

Bug: 77234888
Bug: 73506513
Test: NeuralNetworksTest_static
Change-Id: I4b6281cdd7758054a1a43ac34f39fc9c7bfead9b