Configure the sample drivers with init_rc and vintf_fragments
init_rc (.rc) informs the device how the sample driver process should
be started. This .rc file is automatically included in the device if
the corresponding binary is uploaded (specified by the device's
device.mk file or equivalent). It is uploaded to the device in the
directory /vendor/etc/init/.

vintf_fragments (.xml) informs the device the name and version of the
sample driver service. This .xml file is automatically included in the
device if the corresponding binary is uploaded (specified by the
device's device.mk file or equivalent). It is uploaded to the device
in the directory /vendor/etc/vintf/manifest/. All these names can be
queried on the device with the following command:
    adb shell lshal --types=v -i

Having the vintf_fragments means we do not need to explicitly add each
HAL instance in each device-specific manifest.xml; instead, we only need
to include the corresponding binary in the device.mk file.

Fixes: 120283896
Test: mma
Test: NeuralNetworksTest_static
Test: VtsHalNeuralnetworksV1_*TargetTest
Test: adb shell lshal --types=v -i
Change-Id: Id649b1adb1c6594602750709c9675afc500bc257