Regenerate test codes with the new test generator.
This is a subsequence CL to ag/5064820, including all the re-generated
test codes after the test generator refactory. Please refer to the
previous CL for more description of the test generator.

1. Move cpp examples definition (std::vector<MixedTypedExample> ...)
from test file to example file to support multiple sets of examples.
All the generated files in runtime/test/generated/examples/ and
runtime/test/generated/tests/ are thus affected.

2. Test files in runtime/test/generated/tests/ are now generated by
python script instead of shell script. Modify the commented header.

3. The order of definition for types and operands in generated model
files may be different from the original order, but the new generated
codes are equivalent to the originals, and the new order is tested to be
stable across multiple generations.

4. The order of the operations in generated model files may be different
from the original order, but they are topologically sorted and the order
is stable.

Bug: 114704871

Test: NeuralNetworksTest_static
Change-Id: Idd76394f8e9a61825b175c9c6ab06b239d00ada0