"Lifts" PhiloxRandom from TensorFlow.
This is a minimal subset of PhiloxRandom from:
external/tensorflow/tensorflow/core/lib/random
commit fab5c4425882bcaa374bdf179f4238a616f83372

For use in the implementation of the RANDOM_MULTINOMIAL operation
in order to get the same results as in TensorFlow.
Matching those results will help with testing.

TODO: Refactor the TensorFlow library for use in TFLite and NNAPI
and remove this.

Test: Builds with initial implementation for MULTINOMIAL operation.
Bug: 113562775
Change-Id: Ic9dcca93231b691e4259d43460778c93384e9003