Systrace parser: handle different driver threading
Handling of driver execution time is special-cased to interpret
HIDL execution and notification markers as drivers don't have explicit
tracing (yet?).

This change fixes the case where the execution happens on a different
thread than the notification, which is intended to be possible. This
occurs at least in the sample driver.

Test: ./run_tests.sh
Change-Id: Ib686b60cfcd6a1df38af76d889f1930b9db08181