Implement the following introspection & control APIs:
  - ANeuralNetworksModel_getSupportedOperationsForDevices
  - ANeuralNetworksCompilation_createForDevices

The APIs listed above are added to NDK with corresponding NNAPI runtime
changes. Validation tests and integration tests with mocked CPU driver
are also added. The tests also test APIs to get a device handle and
to query basic information of selected devices.

The changes in DeviceManager are mainly to facilitate testing.

Bug: 112661538
Bug: 112661284
Bug: 111425781
Test: mm
Test: NeuralNetworksTest_static
Change-Id: If7706447ddee993973397765e2eba15845d8bc51