Refactor HAL version validation
Before this change, validateOperation would return the minimal supported
HAL version for the operation, which the caller would compare with the
desired HAL version.

After this change, validateOperation takes the desired HAL version as an
argument and performs the check as part of validation.

This allows for an immutable validation context to be in the
OperationResolver CL http://ag/5327519.

Bug: 114704873
Test: NeuralNetworksTest_static
Change-Id: I253c4c6c50d0304a63b8ecc82c64eb13341317d3