switch to using android-base/file.h instead of android-base/test_utils.h
Reduce the usage footprint of android-base/test_utils.h that do not in
fact need it.  In this case the file really was using test_utils.h
as a way of catching side-effect includes and did not use anything
in test_utils.h.  There is a real possibility of deprecating the
users of test_utils.h when all of them have been rationalized.

Test: compile
Bug: 119313545
Change-Id: I664fb32522d01909c603d7b903475c4e9aea9223