Support @1.2::IPreparedModel::executeSynchronously().
- Implement SamplePreparedModel::executeSynchronously().
- Call executeSynchronously() when available.

Add two more iterations to NeuralNetworksTest, in which
we do not call executeSynchronously() even if it is available.

Add new property debug.nn.syncexec-hal that defaults to 1.
When set to 0, the runtime WILL NOT call executeSynchronously()
even if it is available.

Bug: 119274127

Test: NeuralNetworksTest_static

Change-Id: I6d842519db732bfca057981d7abceebe69499087
(cherry picked from commit 1120420f54c73803a9b3a03c19b34ac3ed197904)