NNAPI Burst -- NDK interface
The NNAPI is introducing the notion of an "Execution Burst" object (or
more simply a "Burst" object), which is similar to an
ANeuralNetworksExecution, but is intended to be reused across multiple
executions and has lower IPC overheads. It achieves this low IPC
overhead by replacing HIDL HwBinder calls with FMQ messages.

This CL exposes the Burst object to the user through the NDK.
Specifically, it adds ANeuralNetworksBurst to NeuralNetworks.h. It declares
functions to create, free, and compute an execution on the Burst in
NeuralNetworks.h, and has a skeletal implementation in
NeuralNetworks.cpp. It exposes these functions in the NDK in
libneuralnetworks.map.txt.

Bug: 119570067
Test: mma
Change-Id: Ic4c562dce3b27c3fd9f7a6234105f475a3969591