Enable sample drivers to provide custom CPU implementations
This change adds new constructors taking IOperationResolver* to SampleDriver
and CpuExecutor.

To provide custom CPU operation implementations the driver should supply an
instance of IOperationResolver to the SampleDriver constructor. That
IOperationResolver instance might delegate some operation lookups to
BuiltinOperationResolver, while returning OperationRegistrations pointing
to custom CPU implementations for other operations.

Note that it is not possible to provide custom implementations for
non-OperationResolver operations.

Bug: 118606713
Test: NeuralNetworksTest_static
Test: NeuralNetworksTest_FibonacciExtension (from change Ibe0fc5356baa909bce8424138bd5cfac9f74648f)
Change-Id: Id6e903e237f41a9834ac4e6d5ce5615316d35176