Add @V1_2::Capabilities to support all non extension operand types.
Performance information in Capabilities is used by the runtime when
it selects the appropriate processor to distribute work to.  Prior to
this CL, Capabilities can only distinguish between float and non-float
data types -- so, for example, float16 and float32 performance is
considered to be the same, and performance for all non-float data types
is considered to be the same.

Also:
- tweak some diagnostics
- fix bug in ::android::nn::wrapper::OperandType and
  ::android::nn::extension_wrapper::OperandType
  copy constructor and assignment operator

Bug: 124041010

Test: NeuralNetworksTest_static
Test: VtsHalNeuralnetworksV1_2TargetTest --hal_service_instance=android.hardware.neuralnetworks@1.2::IDevice/sample-all

Change-Id: If3290542020ac1862bcd94bf3acddbc2e50109a6