Remove tmp cache directory if test succeeds.
Add a new test fixture CacheTokenTest for the original
PartitioningTest.Cache*, that creates a tmp cache dir during SetUp and
removes it during TearDown if test passes.

TestCompilationCaching.CompilationCachingTest/* will now remove the tmp
cache dirs during the TearDown phase if there is no failure in the test.

If the test fails, the dir and cache files are preserved for debugging
purpose in both fixtures.

Test: NeuralNetworksTest_static
Test: Check /data/local/tmp and confirm that
      - Before this CL, hundreds of tmp dirs are preserved after test
      - After this CL
        * dirs are preserved on failure
	* dirs are removed on success

Change-Id: I9599133935660df2b0a26e0c0871ea07ed882f85