Validate DEPTHWISE_CONV_2D outputScale > inputScale * filterScale below v1.2
DEPTHWISE_CONV_2D's constraint "outputScale > inputScale * filterScale" was
relaxed for HAL v1.2, but the appropriate version checks were not put in
place. For more information about this check, refer to the description
of the output operand for DEPTHWISE_CONV_2D in NeuralNetworks.h.

+ Fix of GeneratedTests.depthwise_conv2d_v1_2_quant_output_multiplier_gt_1,
which was not proper (was not testing the case where
'outputScale > inputScale * filterScale')

Bug: 130169097
Test: NeuralNetworkTest_static
Change-Id: I9b39b53f7662638ce9d56189a96dcb05e485d9c2