Allow drivers to fail on compilation or execution.
Even if the model is fully supported, the compilation may still fail,
e.g. each operation is supported, but model is too big (too many
operations and/or too-large constants) for device.

Even if the model is fully supported and the compilation succeeds,
the execution may still fail, e.g. there may be operand shapes that
are unknown until execution time, and at execution time turn out to
be too big.

Bug: 131858477
Test: NeuralNetworksTest_static_fuzzing
Change-Id: I5539b413d5f55cbfdc1d53d90668a6e02fa3eec8