Test compliance along with generated tests.
Generate TEST_AVAILABLE_SINCE for every generated test, with two
exceptions:
* Do not generate for relaxed mode models as relaxed mode won't
affect the compliance check result.
* Do not generate for dynamic output shape tests as dynamic output
shape is a request pattern but not a model pattern.

Most of the model versions are inferred from the directory name, e.g. if
the spec file lives in V1_2 directory then we assume it is available
since V1_2. However there are several exceptions that a certain
operation is under-tested in previous version and more tests are added
in a later version, e.g. V1_1/fully_connected_float_4d_simple.mod.py.
In the case, the version is manually set in spec.

Addtionally, set depthwise_conv_relaxed.mod.py to be in relaxed mode.

The tests are regenerated in the following CL.

Bug: 132090511
Test: NeuralNetworksTest_static
Change-Id: Iaa620b8e5105a4e4a95a62ac306832a0d31ef2a1