Swap resize_bilinear tests' width and height fields
Both the width and height fields in the test are 3, so this is only a
naming change done for clarity and consistency with the APIs.

This CL additionally flips the comment order of "h" and "w" in the
resize_nearest_neighbor test to remain consistent with the comment
scheme in resize_bilinear.

Bug: 131623949
Bug: 130035110
Test: mma
Change-Id: Iade3a1f2ff336195fcd09ffbd27b90043abc7190
Merged-In: Iade3a1f2ff336195fcd09ffbd27b90043abc7190
(cherry picked from commit d1d9cd8486333053d630b41fbdb0b2c669ebddc4)